move filters into writer module
